♻️ Move step-decision logic to helper file
