+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/Onboarding.js
@@ -1,24 +1,14 @@
 import Container from '../../ReusableComponents/Container';
 import { OnboardingHooks } from '../../../data';
-import { getSteps } from './availableSteps';
+
+import { getSteps, getCurrentStep } from './availableSteps';
 import Navigation from './Components/Navigation';
-
-const getCurrentStep = ( requestedStep, steps ) => {
-	const isValidStep = ( step ) =>
-		typeof step === 'number' &&
-		Number.isInteger( step ) &&
-		step >= 0 &&
-		step < steps.length;
-
-	const safeCurrentStep = isValidStep( requestedStep ) ? requestedStep : 0;
-	return steps[ safeCurrentStep ];
-};
 
 const Onboarding = () => {
 	const { step, setStep, setCompleted, flags } = OnboardingHooks.useSteps();
-	const steps = getSteps( flags );
 
-	const CurrentStepComponent = getCurrentStep( step, steps );
+	const Steps = getSteps( flags );
+	const CurrentStepComponent = getCurrentStep( step, Steps );
 
 	return (
 		<>
@@ -26,7 +16,7 @@ const Onboarding = () => {
 				setStep={ setStep }
 				currentStep={ step }
 				setCompleted={ setCompleted }
-				stepperOrder={ steps }
+				stepperOrder={ Steps }
 			/>
 			<Container page="onboarding">
 				<div className="ppcp-r-card">
@@ -34,7 +24,7 @@ const Onboarding = () => {
 						setStep={ setStep }
 						currentStep={ step }
 						setCompleted={ setCompleted }
-						stepperOrder={ steps }
+						stepperOrder={ Steps }
 					/>
 				</div>
 			</Container>
